+++ src/main/java/com/telepathicgrunt/tact/mixin/DesolateDaggerEntityMixin.java
@@ -15,15 +15,13 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class DesolateDaggerEntityMixin {
 
     @ModifyConstant(method = "tick()V",
-            constant = {@Constant(floatValue = 2.0F, ordinal = 0)},
-            remap = false)
+            constant = {@Constant(floatValue = 2.0F, ordinal = 0)})
     private float tact_desolateDaggerAbilityDamage1(float damage) {
         return Config.DESOLATE_DAGGERS_ABILITY_BASE_DAMAGE.get().floatValue();
     }
 
     @ModifyConstant(method = "tick()V",
-            constant = {@Constant(floatValue = 2.0F, ordinal = 1)},
-            remap = false)
+            constant = {@Constant(floatValue = 2.0F, ordinal = 1)})
     private float tact_desolateDaggerAbilityDamage2(float damage) {
         return Config.DESOLATE_DAGGERS_ABILITY_IMPEDING_STAB_DAMAGE.get().floatValue();
     }