+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ### **(V.1.2.2 Changes) (1.20.1 Minecraft)**
 
 #### Configs:
-
+Added configs for Desolate Dagger's ability delay and damage.