+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-### **(V.1.2.2 Changes) (1.20.1 Minecraft)**
+### **(V.1.2.3 Changes) (1.20.1 Minecraft)**
+
+#### Misc:
+Fix startup crash
 
 #### Configs:
 Added configs for Desolate Dagger's ability delay and damage.